+++ src/lib/components/ui/card/card-title.svelte
@@ -10,6 +10,6 @@
 	let { children, ...rest }: Props & SvelteHTMLElements['h2'] = $props();
 </script>
 
-<h2 {...rest} class={cn('mt-6 mb-4 text-2xl font-bold md:text-3xl', rest.class)}>
+<h2 {...rest} class={cn('mt-2 mb-4 text-2xl font-bold md:text-3xl', rest.class)}>
 	{@render children?.()}
 </h2>
